feat: Add deleteOrder mutation and useDeleteOrderMutation
This commit adds the deleteOrder mutation and useDeleteOrderMutation hook to the orderApi in the frontend codebase. The mutation and hook allow the frontend to send a DELETE request to remove an order from the admin dashboard. The deleteOrder mutation includes a body parameter with the order ID to identify the order to be deleted. The useDeleteOrderMutation hook provides a convenient way to use the mutation in components. This feature enhances the admin dashboard by providing the ability to easily delete unwanted orders.