feat: Remove user avatar when deleting a user
This commit modifies the deleteUser function in the authController.js file to remove the user's avatar from the cloudinary storage when deleting a user. If the user has an avatar with a public_id, the removefile function is called to delete the avatar from the cloudinary storage. This enhancement ensures that the user's avatar is properly removed when deleting a user.