chore: Update favicon link in index.html
This commit updates the favicon link in the index.html file. The previous link was pointing to a SVG file, but it has been changed to point to a PNG file named "favicon.png". This change ensures that the correct favicon is displayed in the browser tab and improves the visual representation of the website.