feat: Update favicon link to use new SVG file
This commit updates the favicon link in the index.html file to use a new SVG file named "logogdget.svg". The previous link was pointing to a PNG file, but it has been changed to use the SVG file for better scalability and visual quality. This change ensures that the correct favicon is displayed in the browser tab and improves the visual representation of the website.