feat: Add 'Our Store' page and link to bottom navigation bar
This commit adds a new page called 'Our Store' and a corresponding link to the bottom navigation bar. The 'Our Store' page is implemented in the `OurStore.jsx` file and can be accessed via the '/store' path. This enhancement improves the navigation experience by providing easy access to the store section of the website.Also changing the CSS of the product look.